Refactor Event loading code
