+++ components/dashboard/event-card/index.tsx
@@ -1,12 +1,11 @@
 import * as S from "./styles";
 import { Link } from "components/links";
 import { getResizedImgUrl } from "lib/utils";
-import { PortalEvent } from "lib/portal-types";
 import strings from "content/strings.json";
 import { Route } from "lib/utils";
 import DateTime from "components/datetime";
-import { isEventPast } from "lib/portal-type-utils";
 import { PortalProject } from "lib/airtable/project";
+import { isEventPast, PortalEvent } from "lib/airtable/event";
 
 interface Props {
   event: PortalEvent;
