+++ lib/local-data.ts
@@ -1,8 +1,9 @@
 import fs from "fs";
+import { PortalEvent } from "./airtable/event";
 import { PortalProject } from "./airtable/project";
 import { Field } from "./airtable/skills";
 import { PortalUser } from "./airtable/user";
-import { PortalEvent, PortalOpportunity, PortalPartner } from "./portal-types";
+import { PortalOpportunity, PortalPartner } from "./portal-types";
 
 // The weird signature is here to make the data source type-compatible with the Airtable data source
 function loader<T>(file: string): () => Promise<T[]> {
