+++ lib/utils.ts
@@ -1,7 +1,8 @@
 import { marked } from "marked";
 import { PortalVideo } from "./cedu";
-import { PortalEvent, PortalOpportunity } from "lib/portal-types";
+import { PortalOpportunity } from "lib/portal-types";
 import { PortalProject } from "./airtable/project";
+import { PortalEvent } from "./airtable/event";
 
 /** Create URLs for frequently used routes */
 export const Route = {
