+++ pages/projects/[slug].tsx
@@ -15,10 +15,14 @@ import { Route } from "lib/utils";
 import { Article } from "lib/related-blog-posts";
 import { BlogCard } from "components/cards";
 import EventCard from "components/dashboard/event-card";
-import { compareEventsByTime, isEventPast } from "lib/portal-type-utils";
 import { PortalUser } from "lib/airtable/user";
-import { PortalEvent, PortalOpportunity } from "lib/portal-types";
+import { PortalOpportunity } from "lib/portal-types";
 import { PortalProject } from "lib/airtable/project";
+import {
+  compareEventsByTime,
+  isEventPast,
+  PortalEvent,
+} from "lib/airtable/event";
 
 interface PageProps {
   project: PortalProject;
